Move new semantic types to experimental
