+++ autogoal/kb/_semantics.py
@@ -483,25 +483,6 @@ Tensor3 = Tensor[3, Continuous, Dense]
 Tensor4 = Tensor[4, Continuous, Dense]
 
 
-# Semantics types for audio processing
-class AudioFile(SemanticType):
-    '''Semantic type for wav audio files'''
-    @classmethod
-    def _match(cls, x):
-        try:
-            return path.isfile(x) and x.split('.')[-1] == 'wav'
-        except TypeError:
-            return False
-
-
-class AudioCommand(SemanticType):
-    '''Semantic type for vectors representing audio commands'''
-    @classmethod
-    def _match(cls, x):
-        pass
-
-
-AudioFeatures = Tensor[3, Continuous, Dense]
 
 
 # Finally we define the publicly export classes
@@ -538,9 +519,6 @@ __all__ = [
     "Categorical",
     "Continuous",
     "Discrete",
-    "AudioFile",
-    "AudioCommand",
-    "AudioFeatures"
 ]
